+++ core/src/main/kotlin/com/gitlab/kordlib/core/cache/data/VoiceStateData.kt
@@ -23,7 +23,7 @@ data class VoiceStateData(
     companion object {
         val description = description(VoiceStateData::id)
 
-        fun from(entity: DiscordVoiceState) = with(entity) {
+        fun from(guildId: String?, entity: DiscordVoiceState) = with(entity) {
             VoiceStateData(
                     guildId?.toLong(),
                     channelId?.toLong(),